+++ api/cosmos/orm/module/v1alpha1/module.pulsar.go
@@ -383,8 +383,8 @@ const (
 )
 
 // Module defines the ORM module which adds providers to the app container for
-// module-scoped DB's. In the future it may provide gRPC services for interacting
-// with ORM data.
+// ORM ModuleDB's and in the future will automatically register query
+// services for modules that use the ORM.
 type Module struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache