+++ x/gov/migrations/v4/store.go
@@ -27,10 +27,13 @@ func migrateParams(ctx sdk.Context, storeKey storetypes.StoreKey, legacySubspace
 
 	params := govv1.NewParams(
 		dp.MinDeposit,
+		govv1.DefaultParams().ExpeditedMinDeposit,
 		*dp.MaxDepositPeriod,
 		*vp.VotingPeriod,
+		*govv1.DefaultParams().ExpeditedVotingPeriod,
 		tp.Quorum,
 		tp.Threshold,
+		govv1.DefaultParams().ExpeditedThreshold,
 		tp.VetoThreshold,
 		sdk.ZeroDec().String(),
 		sdk.ZeroDec().String(),