+++ x/gov/simulation/genesis.go
@@ -17,24 +17,41 @@ import (
 
 // Simulation parameter constants
 const (
-	DepositParamsMinDeposit    = "deposit_params_min_deposit"
-	DepositParamsDepositPeriod = "deposit_params_deposit_period"
-	DepositMinInitialRatio     = "deposit_params_min_initial_ratio"
-	VotingParamsVotingPeriod   = "voting_params_voting_period"
-	TallyParamsQuorum          = "tally_params_quorum"
-	TallyParamsThreshold       = "tally_params_threshold"
-	TallyParamsVeto            = "tally_params_veto"
-	ProposalCancelRate         = "proposal_cancel_rate"
+	MinDeposit            = "min_deposit"
+	ExpeditedMinDeposit   = "expedited_min_deposit"
+	DepositPeriod         = "deposit_period"
+	MinInitialRatio       = "min_initial_ratio"
+	VotingPeriod          = "voting_period"
+	ExpeditedVotingPeriod = "expedited_voting_period"
+	Quorum                = "quorum"
+	Threshold             = "threshold"
+	ExpeditedThreshold    = "expedited_threshold"
+	Veto                  = "veto"
+	ProposalCancelRate    = "proposal_cancel_rate"
+
+	// ExpeditedThreshold must be at least as large as the regular Threshold
+	// Therefore, we use this break out point in randomization.
+	tallyNonExpeditedMax = 500
+
+	// Similarly, expedited voting period must be strictly less than the regular
+	// voting period to be valid. Therefore, we use this break out point in randomization.
+	expeditedMaxVotingPeriod = 60 * 60 * 24 * 2
 )
 
-// GenDepositParamsDepositPeriod returns randomized DepositParamsDepositPeriod
-func GenDepositParamsDepositPeriod(r *rand.Rand) time.Duration {
+// GenDepositPeriod returns randomized DepositPeriod
+func GenDepositPeriod(r *rand.Rand) time.Duration {
 	return time.Duration(simulation.RandIntBetween(r, 1, 2*60*60*24*2)) * time.Second
 }
 
-// GenDepositParamsMinDeposit returns randomized DepositParamsMinDeposit
-func GenDepositParamsMinDeposit(r *rand.Rand, bondDenom string) sdk.Coins {
-	return sdk.NewCoins(sdk.NewInt64Coin(bondDenom, int64(simulation.RandIntBetween(r, 1, 1e3))))
+// GenMinDeposit returns randomized MinDeposit
+func GenMinDeposit(r *rand.Rand, bondDenom string) sdk.Coins {
+	return sdk.NewCoins(sdk.NewInt64Coin(bondDenom, int64(simulation.RandIntBetween(r, 1, 1e3/2))))
+}
+
+// GenExpeditedMinDeposit returns randomized ExpeditedMinDeposit
+// It is always greater than GenMinDeposit
+func GenExpeditedMinDeposit(r *rand.Rand, bondDenom string) sdk.Coins {
+	return sdk.NewCoins(sdk.NewInt64Coin(bondDenom, int64(simulation.RandIntBetween(r, 1e3/2, 1e3))))
 }
 
 // GenDepositMinInitialRatio returns randomized DepositMinInitialRatio
@@ -47,23 +64,33 @@ func GenProposalCancelRate(r *rand.Rand) sdk.Dec {
 	return sdk.NewDec(int64(simulation.RandIntBetween(r, 0, 99))).Quo(sdk.NewDec(100))
 }
 
-// GenVotingParamsVotingPeriod returns randomized VotingParamsVotingPeriod
-func GenVotingParamsVotingPeriod(r *rand.Rand) time.Duration {
-	return time.Duration(simulation.RandIntBetween(r, 1, 2*60*60*24*2)) * time.Second
+// GenVotingPeriod returns randomized VotingPeriod
+func GenVotingPeriod(r *rand.Rand) time.Duration {
+	return time.Duration(simulation.RandIntBetween(r, expeditedMaxVotingPeriod, 2*expeditedMaxVotingPeriod)) * time.Second
 }
 
-// GenTallyParamsQuorum returns randomized TallyParamsQuorum
-func GenTallyParamsQuorum(r *rand.Rand) math.LegacyDec {
+// GenExpeditedVotingPeriod randomized ExpeditedVotingPeriod
+func GenExpeditedVotingPeriod(r *rand.Rand) time.Duration {
+	return time.Duration(simulation.RandIntBetween(r, 1, expeditedMaxVotingPeriod)) * time.Second
+}
+
+// GenQuorum returns randomized Quorum
+func GenQuorum(r *rand.Rand) math.LegacyDec {
 	return sdk.NewDecWithPrec(int64(simulation.RandIntBetween(r, 334, 500)), 3)
 }
 
-// GenTallyParamsThreshold returns randomized TallyParamsThreshold
-func GenTallyParamsThreshold(r *rand.Rand) math.LegacyDec {
-	return sdk.NewDecWithPrec(int64(simulation.RandIntBetween(r, 450, 550)), 3)
+// GenThreshold returns randomized Threshold
+func GenThreshold(r *rand.Rand) math.LegacyDec {
+	return sdk.NewDecWithPrec(int64(simulation.RandIntBetween(r, 450, tallyNonExpeditedMax+1)), 3)
 }
 
-// GenTallyParamsVeto returns randomized TallyParamsVeto
-func GenTallyParamsVeto(r *rand.Rand) math.LegacyDec {
+// GenExpeditedThreshold randomized ExpeditedThreshold
+func GenExpeditedThreshold(r *rand.Rand) sdk.Dec {
+	return sdk.NewDecWithPrec(int64(simulation.RandIntBetween(r, tallyNonExpeditedMax, 550)), 3)
+}
+
+// GenVeto returns randomized Veto
+func GenVeto(r *rand.Rand) math.LegacyDec {
 	return sdk.NewDecWithPrec(int64(simulation.RandIntBetween(r, 250, 334)), 3)
 }
 
@@ -73,19 +100,25 @@ func RandomizedGenState(simState *module.SimulationState) {
 
 	var minDeposit sdk.Coins
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, DepositParamsMinDeposit, &minDeposit, simState.Rand,
-		func(r *rand.Rand) { minDeposit = GenDepositParamsMinDeposit(r, simState.BondDenom) },
+		simState.Cdc, MinDeposit, &minDeposit, simState.Rand,
+		func(r *rand.Rand) { minDeposit = GenMinDeposit(r, simState.BondDenom) },
+	)
+
+	var expeditedMinDeposit sdk.Coins
+	simState.AppParams.GetOrGenerate(
+		simState.Cdc, ExpeditedMinDeposit, &expeditedMinDeposit, simState.Rand,
+		func(r *rand.Rand) { expeditedMinDeposit = GenExpeditedMinDeposit(r, simState.BondDenom) },
 	)
 
 	var depositPeriod time.Duration
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, DepositParamsDepositPeriod, &depositPeriod, simState.Rand,
-		func(r *rand.Rand) { depositPeriod = GenDepositParamsDepositPeriod(r) },
+		simState.Cdc, DepositPeriod, &depositPeriod, simState.Rand,
+		func(r *rand.Rand) { depositPeriod = GenDepositPeriod(r) },
 	)
 
 	var minInitialDepositRatio sdk.Dec
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, DepositMinInitialRatio, &minInitialDepositRatio, simState.Rand,
+		simState.Cdc, MinInitialRatio, &minInitialDepositRatio, simState.Rand,
 		func(r *rand.Rand) { minInitialDepositRatio = GenDepositMinInitialDepositRatio(r) },
 	)
 
@@ -97,31 +130,43 @@ func RandomizedGenState(simState *module.SimulationState) {
 
 	var votingPeriod time.Duration
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, VotingParamsVotingPeriod, &votingPeriod, simState.Rand,
-		func(r *rand.Rand) { votingPeriod = GenVotingParamsVotingPeriod(r) },
+		simState.Cdc, VotingPeriod, &votingPeriod, simState.Rand,
+		func(r *rand.Rand) { votingPeriod = GenVotingPeriod(r) },
+	)
+
+	var expeditedVotingPeriod time.Duration
+	simState.AppParams.GetOrGenerate(
+		simState.Cdc, ExpeditedVotingPeriod, &expeditedVotingPeriod, simState.Rand,
+		func(r *rand.Rand) { expeditedVotingPeriod = GenExpeditedVotingPeriod(r) },
 	)
 
 	var quorum sdk.Dec
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, TallyParamsQuorum, &quorum, simState.Rand,
-		func(r *rand.Rand) { quorum = GenTallyParamsQuorum(r) },
+		simState.Cdc, Quorum, &quorum, simState.Rand,
+		func(r *rand.Rand) { quorum = GenQuorum(r) },
 	)
 
 	var threshold sdk.Dec
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, TallyParamsThreshold, &threshold, simState.Rand,
-		func(r *rand.Rand) { threshold = GenTallyParamsThreshold(r) },
+		simState.Cdc, Threshold, &threshold, simState.Rand,
+		func(r *rand.Rand) { threshold = GenThreshold(r) },
+	)
+
+	var expitedVotingThreshold sdk.Dec
+	simState.AppParams.GetOrGenerate(
+		simState.Cdc, ExpeditedThreshold, &expitedVotingThreshold, simState.Rand,
+		func(r *rand.Rand) { expitedVotingThreshold = GenExpeditedThreshold(r) },
 	)
 
 	var veto sdk.Dec
 	simState.AppParams.GetOrGenerate(
-		simState.Cdc, TallyParamsVeto, &veto, simState.Rand,
-		func(r *rand.Rand) { veto = GenTallyParamsVeto(r) },
+		simState.Cdc, Veto, &veto, simState.Rand,
+		func(r *rand.Rand) { veto = GenVeto(r) },
 	)
 
 	govGenesis := v1.NewGenesisState(
 		startingProposalID,
-		v1.NewParams(minDeposit, depositPeriod, votingPeriod, quorum.String(), threshold.String(), veto.String(), minInitialDepositRatio.String(), proposalCancelRate.String(), ""),
+		v1.NewParams(minDeposit, expeditedMinDeposit, depositPeriod, votingPeriod, expeditedVotingPeriod, quorum.String(), threshold.String(), expitedVotingThreshold.String(), veto.String(), minInitialDepositRatio.String(), proposalCancelRate.String(), ""),
 	)
 
 	bz, err := json.MarshalIndent(&govGenesis, "", " ")