+++ x/gov/types/v1/params.go
@@ -11,14 +11,18 @@ import (
 
 // Default period for deposits & voting
 const (
-	DefaultPeriod time.Duration = time.Hour * 24 * 2 // 2 days
+	DefaultPeriod                         time.Duration = time.Hour * 24 * 2 // 2 days
+	DefaultExpeditedPeriod                time.Duration = time.Hour * 24 * 1 // 1 day
+	DefaultMinExpeditedDepositTokensRatio               = 5
 )
 
 // Default governance params
 var (
 	DefaultMinDepositTokens          = sdk.NewInt(10000000)
+	DefaultMinExpeditedDepositTokens = DefaultMinDepositTokens.Mul(math.NewInt(DefaultMinExpeditedDepositTokensRatio))
 	DefaultQuorum                    = sdk.NewDecWithPrec(334, 3)
 	DefaultThreshold                 = sdk.NewDecWithPrec(5, 1)
+	DefaultExpeditedThreshold        = sdk.NewDecWithPrec(667, 3)
 	DefaultVetoThreshold             = sdk.NewDecWithPrec(334, 3)
 	DefaultMinInitialDepositRatio    = sdk.ZeroDec()
 	DefaultProposalCancelRatio       = sdk.MustNewDecFromStr("0.5")
@@ -51,15 +55,18 @@ func NewVotingParams(votingPeriod *time.Duration) VotingParams {
 
 // NewParams creates a new Params instance with given values.
 func NewParams(
-	minDeposit sdk.Coins, maxDepositPeriod time.Duration, votingPeriod time.Duration,
-	quorum, threshold, vetoThreshold, minInitialDepositRatio, proposalCancelRatio, proposalCancelDest string,
+	minDeposit, expeditedminDeposit sdk.Coins, maxDepositPeriod, votingPeriod, expeditedVotingPeriod time.Duration,
+	quorum, threshold, expeditedThreshold, vetoThreshold, minInitialDepositRatio, proposalCancelRatio, proposalCancelDest string,
 ) Params {
 	return Params{
 		MinDeposit:             minDeposit,
+		ExpeditedMinDeposit:    expeditedminDeposit,
 		MaxDepositPeriod:       &maxDepositPeriod,
 		VotingPeriod:           &votingPeriod,
+		ExpeditedVotingPeriod:  &expeditedVotingPeriod,
 		Quorum:                 quorum,
 		Threshold:              threshold,
+		ExpeditedThreshold:     expeditedThreshold,
 		VetoThreshold:          vetoThreshold,
 		MinInitialDepositRatio: minInitialDepositRatio,
 		ProposalCancelRatio:    proposalCancelRatio,
@@ -71,10 +78,13 @@ func NewParams(
 func DefaultParams() Params {
 	return NewParams(
 		sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, DefaultMinDepositTokens)),
+		sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, DefaultMinExpeditedDepositTokens)),
 		DefaultPeriod,
 		DefaultPeriod,
+		DefaultExpeditedPeriod,
 		DefaultQuorum.String(),
 		DefaultThreshold.String(),
+		DefaultExpeditedThreshold.String(),
 		DefaultVetoThreshold.String(),
 		DefaultMinInitialDepositRatio.String(),
 		DefaultProposalCancelRatio.String(),
@@ -84,8 +94,15 @@ func DefaultParams() Params {
 
 // ValidateBasic performs basic validation on governance parameters.
 func (p Params) ValidateBasic() error {
-	if minDeposit := sdk.Coins(p.MinDeposit); minDeposit.Empty() || !minDeposit.IsValid() {
+	minDeposit := sdk.Coins(p.MinDeposit)
+	if minDeposit.Empty() || !minDeposit.IsValid() {
 		return fmt.Errorf("invalid minimum deposit: %s", minDeposit)
+	}
+
+	if minExpeditedDeposit := sdk.Coins(p.ExpeditedMinDeposit); minExpeditedDeposit.Empty() || !minExpeditedDeposit.IsValid() {
+		return fmt.Errorf("invalid expedited minimum deposit: %s", minExpeditedDeposit)
+	} else if minExpeditedDeposit.IsAllLTE(minDeposit) {
+		return fmt.Errorf("expedited minimum deposit must be greater than minimum deposit: %s", minExpeditedDeposit)
 	}
 
 	if p.MaxDepositPeriod == nil {
@@ -118,6 +135,20 @@ func (p Params) ValidateBasic() error {
 		return fmt.Errorf("vote threshold too large: %s", threshold)
 	}
 
+	expeditedThreshold, err := sdk.NewDecFromStr(p.ExpeditedThreshold)
+	if err != nil {
+		return fmt.Errorf("invalid expedited threshold string: %w", err)
+	}
+	if !threshold.IsPositive() {
+		return fmt.Errorf("expedited vote threshold must be positive: %s", threshold)
+	}
+	if threshold.GT(math.LegacyOneDec()) {
+		return fmt.Errorf("expedited vote threshold too large: %s", threshold)
+	}
+	if expeditedThreshold.LTE(threshold) {
+		return fmt.Errorf("expedited vote threshold %s, must be greater than the regular threshold %s", expeditedThreshold, threshold)
+	}
+
 	vetoThreshold, err := sdk.NewDecFromStr(p.VetoThreshold)
 	if err != nil {
 		return fmt.Errorf("invalid vetoThreshold string: %w", err)
@@ -132,9 +163,18 @@ func (p Params) ValidateBasic() error {
 	if p.VotingPeriod == nil {
 		return fmt.Errorf("voting period must not be nil: %d", p.VotingPeriod)
 	}
-
 	if p.VotingPeriod.Seconds() <= 0 {
 		return fmt.Errorf("voting period must be positive: %s", p.VotingPeriod)
+	}
+
+	if p.ExpeditedVotingPeriod == nil {
+		return fmt.Errorf("expedited voting period must not be nil: %d", p.ExpeditedVotingPeriod)
+	}
+	if p.ExpeditedVotingPeriod.Seconds() <= 0 {
+		return fmt.Errorf("expedited voting period must be positive: %s", p.ExpeditedVotingPeriod)
+	}
+	if p.ExpeditedVotingPeriod.Seconds() >= p.VotingPeriod.Seconds() {
+		return fmt.Errorf("expedited voting period %s must be strictly less that the regular voting period %s", p.ExpeditedVotingPeriod, p.VotingPeriod)
 	}
 
 	minInitialDepositRatio, err := sdk.NewDecFromStr(p.MinInitialDepositRatio)